third_party/blink/web_tests/virtual/keyboard-focusable-scrollers-opt-out/shadow-dom/focus-navigation/focus-navigation-scroller-interactive-child-expected.txt

This is a testharness.js-based test.
[FAIL] While focusing a keyboard-focusable scroller, adding interactive content should not cancel focusability
  assert_equals: scroller should be keyboard focusable expected Element node <div id="scroller" style="overflow:scroll; width:50px; he... but got Element node <button id="end">END</button>
Harness: the test ran to completion.
